example/README.md: Make the tutorial checkable
We instrument the tutorial in a way, that all it’s instructions can be
run by `mdsh`. This makes it possible to add it to CI, preventing it
to go out of date when we make changes to the `lorri` tool or any
dependencies.

The tutorial should work flawlessly under any circumstance, otherwise
we lose possible users before they even get to use lorri.