builder/BuildResults: restrict `named_drvs` to just two fields
The nix code in `./src/logged_evaluation.nix` only ever prints two
derivations, but `named_drvs` was generalized to a HashMap which
supports any amount of derivations.

We are better off with a struct in this case, because we can fail
early if we don’t find one of the expected derivations in the nix
output (or find one more than once). It also acts as documentation of
what exactly we expect from the output of `logged_evaluation.nix`.